Still still unfinished changes
I'm 60% through transferring the information from the old flunkd.com
site to this file system.

There are a lot of problems that need to be ironed out. But my first
goal is to get it out there for students to use and get used to so they
can keep offering feedback, and also get a sense of progress from us.

In the next few days we need to add search, finish transferring the
information, and exchange the current CSS used for the extra information
pop-ups for JavaScript so that the latest updates for Chrome and Firefox
can display them.