Finished transferring files from old flunkd
I got all of the files transferred, and fixed the main bugs I found.
Things are pretty consistent between the various calculus topics now,
and the YouTube videos are obeying their z-indexes.

We need to add search functionality and probably change the
hover/mouseover boxes to just mouseover boxes. Then we can put it up for
students to use while we work on kickstarter and other programming.