Major work with the Lesson Builder, and other things
Spent 24 hours straight adding the buttons and figuring out how forms
work with Java. Also figured out the CSS for text explanations for the
lesson builder. Worked on a photo editing API that I almost got set up.
Lots to do, but a lot of the painful and stupid problems are over. I
know what to do from here.