Finally finished the image editing API
It took me probably 20 hours total, and I just realized that it isn't a
very good editing program. Darn. Well I learned how to save images when
they are passed as parameters, so that's good. If I find a really great
API in the future I know what to do.