Finished the basic Lesson Editor
Not secure, but basically functional enough for a friend to use. I added
"Insert Graph." The open source software I was using to generate graphs
before doesn't allow people to access the image they generate through
the URL. But I needed to in order to send to the Fotoflexer API. So I
downloaded the open source program, and he said "you don't have to deny
access to the image through the URL, but I recommend it." So I did a
google search for the specific name of his software, and I found 2
resources that use it and have image URL access enabled!!!

I made the saving images thing better, so now it asks if you want to
save over images.

There is a lot to do still. I barely understand what I've put together.
It could be made more efficient and organized. Now I need to get the
general web styling ready for the Linear Algebra class to use and also
my Modern Physics class.