Lesson Builder "Video" button added
I'm going to use this to re-do all of the video lessons in Calculus
later this semester. Video lessons are going to use the same CSS as Text
Lessons, because in the long run, they might be combined. Also I wanted
to use this while I wrote stuff for Physics 222 in the coming weeks.