Edited a little more with the new design(Design#3)
Some study found that menu transitions should have .5s delay and .1s
transition to optimize results with the users. So that's what I did to
the topbannerContainer in the Chain Rule text lesson.

Next, I need to add the bottom banner, and then make the Physics 222
page, then an adaptive learning page and a feedback page. Then get ready
to add a bunch of 222 files.