Adjusted many things.
I didn't plan on spending this much time on it.

I decided that the way we should create lessons is to make both a right
page and left page text of each lesson, and then put them both in a
folder that has the person's account name on it for when we want to
access it specifically.

One question is, how will we know which lessons need to be rated still?
When they are freshly created, maybe we create them to two locations
until they've been rated enough, then the temporary location is
eliminated and its ratings go to the other lesson. The other lesson's
location will be stored in the temporary file. Maybe there can be a
third file in each person's explanation's folder that keeps the
information available. Maybe.

I fixed a few stuffs. I need to prepare to upload a ton of 222 files.
Should I create them dynamically? Maybe I don't have time for that yet.