I changed many things.
Wrote a bunch of stuff, wasted 3 hours on general relativity because
file history wasn't turned on on Windows 8... But I really need to learn
about JSP and Java and Tomcat more so I don't have to be constantly
putting everything together with duct tape. It needs more elegance I
think.