Practice Problem Picker
Finally got it to work. It chooses the problem with the closest
difficulty level to the preference.

This is a very duct tape type solution. Problems with the same
difficulty are only determined by alphabetical order. In the future
there should be a database that keeps track of the name of each problem,
its difficulty, and its popularity.