Video lessons ready to copy
Check LengthContraction for L.txt and R.txt then copy it to all the
video lessons on the site.

Someday the video lesson.jsp things will be identical to TextLesson.jsp
things because whether it's including a video or a text lesson should
depend on the variables describing the file rather than the name.