30 hidden boxes now, 7 calc topics Latex'ed
Maybe I could call my website "learnidy."

I just saw a 1 min video on YouTube that just came out in December that
predicts the future of education, and it is involved with the main idea
of my site. Classrooms will understand which students are visual
learners and give them visual content. Score! We thought of it first.
The future of education is actually really great.