Got to the end of Calculus 1
Next I think I'm going to style the practice problems and add those,
then tell Dr. Wadley about it, then finish the other 21 or so subjects
at the first half of the course.

Also, I added a "Future of Education" folder just so that I could have a
place to compile my thoughts and videos.