Made the student side problems look normaler
The correct difficulty leve parameter was being passed, and before it
put the slider in the right place it would change the
"displayDifficulty" number. No idea why. But it turns out the slider's
default value is 6.

Still need to brush up the style and then make the problem creator look
good.