A new era of editing has begun
I downloaded "Alway Sync", and now I use the webdisc thing from CPanel
to directly change things on the website. It only takes 3 seconds to
update the changes, which is much better than the super long process I
was using before. Then, I use Alway Sync after I'm done for the day and
then commit the changes that have been synced. This way, I'll also be
able to preserve the stuff I come up with dynamically.

If I need to make a really fast change, the Tomcat deploying ROOT.war
stuff is the way to go, because syncing takes 5 minutes.

Also, I need to find a solution to the MathJax stuff on the Problem
Creator. I have a "$" inside of strings in JS, and it doesn't like it. I
have to change them to \( and \) instead.