Practice Problems now work with $ for mathjax
After 2-3 hours, I realized that in PracticeProblemsHead.jsp I had left
a <script> tag unclosed, and the next script was the one that happened
to replace "\(" with "$" for inline equations for mathjax. It was so
subtle I thought that it could have been from any kind of cause.

But it works now.