More Java, jstl, Course Creator
Course Creator took 15 hours and is not extremely great, but if you go
in knowing exactly what you want to create, it works fine.

I took an hour and made the LessonBuilder have the capability of saving
directly to the lesson folders. I need to add the log still so I have a
record of stuff.

I guess I don't really have anything else to do with learningflare.com.
I'll just wait for shareacademy to become available, and then move the
shareacademy.learningflare.com stuff over. In the meantime I can build
the next stage on that subdomain. The next stage is going to be amazing.